fix: moved isValidCountryCode to shared utils
